requirements.txt
APScheduler==3.6.3
black==23.1.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
flake8==6.0.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pyaes==1.6.1
pycodestyle==2.10.0
pyflakes==3.0.1
Pyrogram==2.0.102
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tomli==2.0.1
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
